refactor(persistence):renaming service, replacing mapper factory with enum extension
